requirements.txt
Django==1.9.5
django-appconf==1.0.2
django-compressor==2.0
django-selenium==0.9.8
djangorestframework==3.3.3
rcssmin==1.0.6
rjsmin==1.0.12
